Look up the owner for an event

+++ src/ApiService/Controllers/EventsController.cs
@@ -24,7 +24,7 @@ namespace Calendar.ApiService.Controllers
         /// Gets data for a given event by its ID.
         /// </summary>
         [HttpGet("{id}")]
-        public async Task<IActionResult> GetEventAsync(int id)
+        public async Task<IActionResult> GetEventAsync(long id)
         {
             var event_ = await eventsProvider.GetEventAsync(id);
             if (event_ == null)
