+++ src/ObjectModel/DataProviders/MemoryEventsProvider.cs
@@ -6,7 +6,7 @@ namespace Calendar.ObjectModel.DataProviders
 {
     public sealed class MemoryEventsProvider : IEventsProvider
     {
-        public Task<Event?> GetEventAsync(int id)
+        public Task<Event?> GetEventAsync(long id)
         {
             return Task.FromResult<Event?>(new Event
             {
